Added a LtLFormula and Batch AP class to store the meaningful data from the speific task for the robot fleet that is conveyed through the ltl formula. I chose to store this in the buch automaton class to allow for smooth transitions between querying data from both the buchi automaton and the ltl formula. Began to write and Debug the Tree search algorithm to make a HMRTA Plan for the specific task laid out in the LTL. Still need to implement better visualizations for the Buchi automaton, the environment, and the Multirobot system

+++ Task Batch Planning Decision Tree/Environment/GridWorld.cpp
@@ -3,7 +3,7 @@
 /**
  * GridWorld - Constructor
  */
-GridWorld::GridWorld(uint32_t w, uint32_t h): width(w), height(h) {
+GridWorld::GridWorld(uint16_t w, uint16_t h): width(w), height(h) {
     // Initialize costmap with all free cells (0)
     costmap.resize(height, std::vector<uint8_t>(width, 0));
 }
